Помощь друга- доработка
