+++ js/main.js
@@ -1,6 +1,7 @@
 import {getRandomIntInclusive, getRandomArbitrary} from './util.js';
 import {TITLES, TYPES, CHECKINS, CHECKOUTS, FEATURES, DESCRIPTIONS, PHOTOS} from './data.js';
-import {createMarkers} from './form.js';
+import {createMarkers} from './map.js';
+import './form.js';
 
 const getRandomArrayElement = (elements) => {
   return elements[getRandomIntInclusive(0, elements.length - 1)];
